# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Mavi API
  version: 0.1.0
  description: |
    JSON CRUD for candidates, their profiles, roles and matches. Every request
    outside `/health` names a persona with the `X-Role` header (`talent`,
    `employer` or `ops`) and may identify itself with `X-Actor` (the candidate
    id for talent, an email for ops). A missing or unknown role is a 401; a
    known role calling an operation it may not use is a 403.

    `PUT` on candidates and roles replaces the fields you send and keeps the
    rest; an explicit `null` clears an optional field and is a 422 on a
    required one. `PUT …/profile` replaces the whole profile. Taxonomy fields
    (certifications, software) accept free text and are stored as canonical
    ids from infra/taxonomy.json; an unknown value is a 422.
servers:
  - url: http://localhost:8080
security:
  - role: []
...
paths:
...
  /candidates/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [candidates]
      operationId: getCandidate
      summary: Fetch a candidate
      description: Talent may only read their own record (`X-Actor` = id); another id is a 404.
      x-roles: [talent, ops]
      responses:
        "200":
          description: The candidate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Candidate" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [candidates]
      operationId: updateCandidate
      summary: Update a candidate
      description: Replaces the fields sent and keeps the rest. Talent may only update their own record.
      x-roles: [talent, ops]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CandidateInput" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Candidate" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      tags: [candidates]
      operationId: deleteCandidate
      summary: Delete a candidate
      x-roles: [ops]
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /candidates/{id}/resume:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [profiles]
      operationId: uploadResume
      summary: Upload a resume and queue its parsing
      description: |
        Talent intake. The request body is the file itself (not multipart): a
        PDF, or a DOCX. What the file is is read from its content, not from
        `Content-Type`. Its text is extracted in the request and stored as the
        candidate's `resume_text`, and a `parse_resume` job is queued; the
        response is that job, whose `id` is what to poll. The worker sends the
        text to the AI service's `/parse-resume`, writes the result as the
        candidate's profile (replacing any profile already there) and embeds
        it. Poll `GET /candidates/{id}/resume/job` (or, as ops,
        `GET /jobs/{id}`) until `status` is `succeeded`, then read
        `GET /candidates/{id}/profile`.

        Uploading again while the job is still queued replaces the text and
        returns the same job. Talent may only upload to their own record.
      x-roles: [talent, ops]
      requestBody:
        required: true
        description: The file itself, at most 5 MB and 10 pages.
        content:
          application/pdf:
            schema: { type: string, format: binary }
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        "202":
          description: Text extracted and stored; the parse is queued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413":
          description: The file is larger than 5 MB.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "415":
          description: The file is neither a PDF nor a DOCX.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: The file is empty, damaged, password-protected, has no text, or has too many pages or too much text; `error` says which.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: The AI service, which reads the file, is unavailable. Nothing was stored; try again.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /candidates/{id}/resume/job:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [profiles]
      operationId: getResumeJob
      summary: Status of the candidate's most recent resume parse
      description: |
        The newest `parse_resume` job for the candidate: the same object
        `GET /jobs/{id}` returns, readable by the talent who uploaded. The
        profile is ready when `status` is `succeeded`; on `failed`,
        `last_error` says why. A 404 when no resume was ever uploaded.
      x-roles: [talent, ops]
      responses:
        "200":
          description: The job.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    role:
      type: apiKey
      in: header
      name: X-Role
      description: A `Persona` (`talent`, `employer` or `ops`). The `mavi_role` cookie is a session fallback.
    actor:
      type: apiKey
      in: header
      name: X-Actor
      description: >-
        The candidate id for talent (required for talent to read anything), an email for ops (required for
        the review operations, whose audit trail records it).
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
...
  responses:
    BadRequest:
      description: Malformed JSON or an unknown field.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or unknown `X-Role`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: >-
        The role may not call this operation, a talent request is missing its candidate id, or a review
        operation is missing its reviewer (`X-Actor`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such row, or one the caller may not see.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: >-
        A unique constraint (duplicate email, duplicate role/candidate pair), a delete blocked by history, or a
        review operation the role's matches do not allow as they stand (`error` says why; nothing was changed).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationFailed:
      description: One or more fields are invalid; `fields` maps each to a message.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Id:
      type: string
      description: A UUID, as text.
      example: 11111111-0000-0000-0000-000000000001
    JobId:
      type: integer
      format: int64
      description: A job's id. Jobs are numbered from a sequence, not UUIDs.
      example: 42
...
    JSONObject:
      type: object
      description: A free-form JSON object stored as JSONB.
      additionalProperties: true
      x-go-type: RawJSON
    CandidateStatus:
      type: string
      enum: [active, archived]
...
    JobStatus:
      type: string
      description: "`queued` → `running` → `succeeded`, or `running` → `queued` (retry) → … → `failed`."
      enum: [queued, running, succeeded, failed]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: validation failed
        fields:
          type: object
          description: Per-field messages on a 422.
          additionalProperties: { type: string }
...
    Candidate:
      type: object
      required: [id, full_name, email, phone, location, resume_text, source, status, created_at, updated_at]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        full_name: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        location: { type: string, nullable: true }
        resume_text: { type: string }
        source:
          type: string
          nullable: true
          description: Where the record came from; `self` when talent created it.
        status: { $ref: "#/components/schemas/CandidateStatus" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CandidateInput:
      type: object
      description: |
        Body for creating or updating a candidate. Every field is optional on
        the wire so a `PUT` can send only what changes; `full_name` is required
        when creating and may not be nulled. `null` clears an optional field.
      properties:
        full_name: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        location: { type: string, nullable: true }
        resume_text:
          type: string
          nullable: true
          description: "`null` clears it to an empty string."
        source: { type: string, nullable: true }
        status:
          allOf: [{ $ref: "#/components/schemas/CandidateStatus" }]
          description: Defaults to `active` on create.

    # -- profiles -------------------------------------------------------------
...
    Job:
      type: object
      description: |
        One background job in the Postgres-backed queue (the `jobs` table). A
        worker claims a `queued` job whose `run_at` has passed, runs the handler
        for its `kind`, and marks it `succeeded`. A failing attempt puts it back
        to `queued` with `run_at` pushed out by a backoff until `attempts`
        reaches `max_attempts`, when it lands in `failed` with `last_error` set.
      required:
        [id, kind, payload, status, priority, run_at, attempts, max_attempts, last_error, worker, started_at,
         finished_at, created_at, updated_at]
      properties:
        id: { $ref: "#/components/schemas/JobId" }
        kind:
          type: string
          description: Names the handler that runs it, e.g. `embed_role`, `parse_resume` or `match_role`.
        payload:
          allOf: [{ $ref: "#/components/schemas/JSONObject" }]
          description: 'Handler-specific input, e.g. `{"role_id": "…"}`.'
        status: { $ref: "#/components/schemas/JobStatus" }
        priority:
          type: integer
          description: Higher runs first.
        run_at:
          type: string
          format: date-time
          description: Not before this time. Retry backoff pushes it into the future.
        attempts:
          type: integer
          description: Claims so far, including one in progress.
        max_attempts:
          type: integer
          minimum: 1
        last_error:
          type: string
          nullable: true
          description: The most recent failed attempt's error; kept on a `failed` job, cleared on success.
        worker:
          type: string
          nullable: true
          description: Id of the worker that last claimed it.
        started_at:
          type: string
          format: date-time
          nullable: true
          description: When the current or most recent attempt was claimed.
        finished_at:
          type: string
          format: date-time
          nullable: true
          description: Set when the job reached `succeeded` or `failed`.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...